sdd/bmad/backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.20  # Required for file upload

# LangChain and AI
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5  # For LLM integration

# Vector Database
chromadb>=0.4.24

# SQL Parser
sqlparse>=0.4.4

# Configuration and Environment
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
